requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
psycopg2-binary>=2.9
pydantic>=2.0
pydantic-settings>=2.0
